requirements.txt
# Core framework
fastapi==0.115.6
uvicorn[standard]==0.32.1

# Cryptography
pycryptodome==3.21.0

# Validation
pydantic==2.10.4
